+++ src/collectors/mod.rs
@@ -1,5 +1,6 @@
 pub mod diskstats;
 #[cfg(test)]
 pub mod iostat;
+pub mod md_sysfs;
 pub mod raid;
 pub mod smart;